feat: Add VS Code extension for Cursor\Trae\Antigravity usage tracking with team collaboration Initialize VS Code extension with usage monitoring for Cursor and Trae, including database setup, status bar display, team server integration, API key management, and web interface for collaborative tracking.
