+++ CodingUsage/README.md
@@ -1,15 +1,21 @@
 # Coding Usage
 
-No need to manually extract Cookies - monitor your Cursor/Trae AI usage, with support for remote data delivery to your self-hosted server.
-Current Cursor usage rules: Basic $20 API usage + official Bonus usage provided by Cursor (currently $25) + Auto usage (currently $150). When the official statistics show "You've hit your usage limit", it only indicates that the basic $20 API usage has been exhausted (this is currently a Cursor statistics bug that incorrectly includes Auto usage in the count).
+Monitor your Cursor/Trae/Antigravity AI usage, supporting multi-account local monitoring with optional data delivery to self-hosted servers.
+
+Current usage rules:
+- **Cursor**: Base $20 API usage + Official bonus ($25 currently) + Auto usage ($150 currently)
+- **Trae**: Monitor API quota usage provided by the official platform
+- **Antigravity**: Monitor quotas and reset times for Claude 4.5, Gemini 3 Pro/Flash, and other models
 
 **English Version | [中文版](README_CN.md)**
 
 <div align="left">
-  <img src="https://raw.githubusercontent.com/lasoons/CodingUsage/refs/heads/main/CodingUsage/img/statebar.png" alt="Status Bar Configuration" width="400">
+  <img src="https://raw.githubusercontent.com/lasoons/CodingUsage/refs/heads/main/CodingUsage/img/statebar_cursor.png" alt="Cursor Status Bar" width="400">
+  <img src="https://raw.githubusercontent.com/lasoons/CodingUsage/refs/heads/main/CodingUsage/img/statebar_trae.png" alt="Trae Status Bar" width="400">
+  <img src="https://raw.githubusercontent.com/lasoons/CodingUsage/refs/heads/main/CodingUsage/img/statebar_antigravity.png" alt="Antigravity Status Bar" width="400">
 </div>
 
-Currently supports [Cursor](cursor:extension/whyuds.coding-usage) and [Trae International](trae:extension/whyuds.coding-usage)
+Currently supports [Cursor](cursor:extension/whyuds.coding-usage), [Trae International](trae:extension/whyuds.coding-usage), and **Antigravity**
 
 ## Configuration Demo
 
@@ -20,11 +26,12 @@ Currently supports [Cursor](cursor:extension/whyuds.coding-usage) and [Trae Inte
 
 ## Features
 
-- **Easy Setup**: Browser extension automatically fetches tokens from the official website and copies to clipboard, IDE extension automatically reads and configures from clipboard
-- **Real-time Updates**: Monitors local AI conversations every 10 seconds, fetches official usage API only when changes are detected
+- **Multi-Platform Support**: One extension supports Cursor, Trae, and Antigravity monitoring with automatic display switching
+- **Zero-Configuration Monitoring**: Antigravity supports automatic local process detection and token retrieval without manual input
+- **Real-time Updates**: Monitor local conversations or processes every 10 seconds/minute, only requesting official API when changes are detected
 - **Team Collaboration**: Optional team server integration for shared usage tracking
-- **Auto Discovery**: Automatically finds and configures available team servers, currently deployed to [Demo Server](http://115.190.183.157:3000/)
-- **Detailed Display**: Comprehensive tooltips with usage details, progress bars, and billing cycle information
+- **Auto-Discovery**: Automatically finds and configures available team servers, currently deployed to [Demo Server](http://115.190.183.157:3000/)
+- **Usage Display**: Detailed tooltips with usage breakdown, progress bars, and billing cycle information
 
 ## Quick Start
 
@@ -41,11 +48,11 @@ Currently supports [Cursor](cursor:extension/whyuds.coding-usage) and [Trae Inte
 
 <p align="center"><em>Double-click Status Bar Configuration | Open QuickPick Menu</em></p>
 
-### 3. Install Browser Extension from QuickPick Menu
+### 3. Install Browser Extension via QuickPick Menu
 
 ### 4. Navigate to Official Website via Browser Extension or QuickPick
 
-### 5. Return and Automatically Configure SessionToken
+### 5. Return and Auto-configure SessionToken
 
 ## Team Features (Optional)
 
@@ -75,26 +82,26 @@ Configure team server URL to enable:
 ## FAQ
 
 #### Why does Cursor Pro subscription show a total of $45?
-Currently, Pro total usage includes: $20 API fixed + $25 Bonus, Auto billing: $150
+Currently, the Pro total usage consists of: Fixed $20 API billing + $25 Bonus, Auto billing: $150
 
-#### How often is the data updated?
-The extension checks the local database every 10 seconds and only calls the official API when changes are detected, ensuring real-time updates while avoiding frequent requests.
+#### How often is data updated?
+The extension checks local database changes every 10 seconds and only calls the official API when changes are detected, ensuring real-time updates while avoiding frequent requests.
 
-#### Is a team server required?
+#### Is the team server required?
 No. The team server is an optional feature for team collaboration, historical data tracking, and personal multi-account tracking.
-The current configured server is only for public demonstration purposes and is disabled by default. It's recommended to deploy your own team server on an internal network according to the submission protocol.
+The current configured server is only for public demonstration, disabled by default. It's recommended to deploy your own team server internally according to the delivery protocol.
 
-#### Team Mode Data Submission Format
+#### Data format for team mode delivery
 
 ```json
 {"client_token":"ck_eb33d6fb4d5b541d28a0d042b0e4ba56","email":"aisrv0615@qiyi.com","expire_time":1767060076000,"membership_type":"pro","api_spend":2002,"api_limit":4500,"auto_spend":0,"auto_limit":15000,"host":"IQ275CG42123NJ","platform":"win32","app_name":"Cursor"}
 ```
 
-#### Is my token secure?
-Tokens are stored locally only. The extension does not send your token to any server except the official API.
+#### Are my tokens safe?
+Tokens are only stored locally. The extension does not send your tokens to any server except the official API.
 
 ---
 
 ## Support
 
-For issues or suggestions, please submit an [Issue](https://github.com/lasoons/CodingUsage/issues) or [Pull Request](https://github.com/lasoons/CodingUsage/pulls).
+For questions or suggestions, please submit an [Issue](https://github.com/lasoons/CodingUsage/issues) or [Pull Request](https://github.com/lasoons/CodingUsage/pulls).
