+++ CodingUsage/src/common/constants.ts
@@ -1,0 +1,14 @@
+export const APP_NAME = 'Coding Usage';
+export const DOUBLE_CLICK_DELAY = 300;
+export const MAX_RETRY_COUNT = 3;
+export const RETRY_DELAY = 1000;
+export const FETCH_TIMEOUT = 30000;
+export const API_TIMEOUT = 5000;
+
+// Trae 只保留默认 host，仅在 storage.json 不可用时作为回退
+export const TRAE_DEFAULT_HOST = 'https://api-sg-central.trae.ai';
+
+export const CURSOR_API_BASE_URL = 'https://cursor.com/api';
+
+
+
