+++ CodingUsage/src/extension.ts
@@ -1,1451 +1,39 @@
 import * as vscode from 'vscode';
-import * as os from 'os';
-import * as path from 'path';
-import * as fs from 'fs-extra';
-import * as crypto from 'crypto';
-import initSqlJs from 'sql.js';
-
 import {
-  disposeOutputChannel,
-  getOutputChannel,
   logWithTime,
-  formatTimestamp,
-  formatTimeWithoutYear,
-  getAdditionalSessionTokens,
-  isRetryableError,
   getAppType,
-  getAppDisplayName,
-  getConfig,
-  getClientApiKey,
-  setClientApiKey,
-  getTeamServerUrl,
-  getClipboardTokenPattern,
-  getDbMonitorKey,
   getDashboardUrl,
+  getClientApiKey,
+  getTeamServerUrl,
+  getConfig,
   isReportingEnabled
-} from './utils';
-import {
-  getApiService,
-  UsageSummaryResponse,
-  BillingCycleResponse,
-  AggregatedUsageResponse,
-  TraeApiResponse,
-  TraeEntitlementPack
-} from './apiService';
-import { ServerDiscovery, TeamServerClient, PingManager, ApiKeyGenerator } from './teamServerClient';
+} from './common/utils';
+import { APP_NAME } from './common/constants';
+import { IUsageProvider } from './common/types';
+import { CursorProvider } from './cursor/cursorProvider';
+import { TraeProvider } from './trae/traeProvider';
+import { AntigravityProvider } from './antigravity/antigravityProvider';
+import { DbMonitor, ClipboardMonitor } from './common/monitors';
+import { ServerDiscovery, TeamServerClient, PingManager } from './teamServerClient';
 
-// ==================== 常量定义 ====================
-const APP_NAME = getAppDisplayName();
-const APP_TYPE = getAppType();
-const DOUBLE_CLICK_DELAY = 300;
-const MAX_RETRY_COUNT = 3;
-const RETRY_DELAY = 1000;
-const FETCH_TIMEOUT = 30000; // 30秒超时
-
-// ==================== Token 自动检测 ====================
-function getGlobalStorageDbPath(): string {
-  const platform = os.platform();
-  const homeDir = os.homedir();
-  const appFolderName = APP_TYPE === 'trae' ? 'Trae' : 'Cursor';
-
-  switch (platform) {
-    case 'win32': {
-      const appData = process.env.APPDATA || path.join(homeDir, 'AppData', 'Roaming');
-      return path.join(appData, appFolderName, 'User', 'globalStorage', 'state.vscdb');
-    }
-    case 'darwin':
-      return path.join(homeDir, 'Library', 'Application Support', appFolderName, 'User', 'globalStorage', 'state.vscdb');
-    default:
-      return path.join(homeDir, '.config', appFolderName, 'User', 'globalStorage', 'state.vscdb');
-  }
-}
-
-async function readAccessTokenFromDb(context: vscode.ExtensionContext): Promise<string | null> {
-  try {
-    const wasmPath = vscode.Uri.joinPath(context.extensionUri, 'out', 'sql-wasm.wasm').fsPath;
-    const dbPath = getGlobalStorageDbPath();
-
-    if (!await fs.pathExists(dbPath)) {
-      logWithTime(`数据库文件不存在: ${dbPath}`);
-      return null;
-    }
-
-    const SQL = await initSqlJs({ locateFile: () => wasmPath });
-    const fileBuffer = await fs.readFile(dbPath);
-    const db = new SQL.Database(fileBuffer);
-    const res = db.exec(`SELECT value FROM ItemTable WHERE key = 'cursorAuth/accessToken';`);
-    db.close();
-
-    if (res && res.length > 0 && res[0].values && res[0].values.length > 0) {
-      const val = res[0].values[0][0];
-      return typeof val === 'string' ? val : null;
-    }
-    return null;
-  } catch (error) {
-    logWithTime(`读取 accessToken 失败: ${error}`);
-    return null;
-  }
-}
-
-function constructSessionToken(accessToken: string): string | null {
-  try {
-    // JWT 格式: header.payload.signature
-    const parts = accessToken.split('.');
-    if (parts.length !== 3) {
-      logWithTime('accessToken 不是有效的 JWT 格式');
-      return null;
-    }
-
-    // 解码 payload (Base64)
-    const payload = JSON.parse(Buffer.from(parts[1], 'base64').toString('utf-8'));
-
-    // sub 格式: auth0|user_01K8VZM8H7ACG6AXT48FTT5089
-    const sub = payload.sub;
-    if (!sub || !sub.includes('|')) {
-      logWithTime(`JWT sub 字段格式不正确: ${sub}`);
-      return null;
-    }
-
-    // 提取 user_XXXXX 部分
-    const userId = sub.split('|')[1];
-
-    // 构造 session token: user_ID::JWT_TOKEN (URL 编码 :: 为 %3A%3A)
-    return `${userId}%3A%3A${accessToken}`;
-  } catch (error) {
-    logWithTime(`解析 JWT 失败: ${error}`);
-    return null;
-  }
-}
-
-async function readCachedEmailFromDb(context: vscode.ExtensionContext): Promise<string | null> {
-  try {
-    const wasmPath = vscode.Uri.joinPath(context.extensionUri, 'out', 'sql-wasm.wasm').fsPath;
-    const dbPath = getGlobalStorageDbPath();
-
-    if (!await fs.pathExists(dbPath)) {
-      return null;
-    }
-
-    const SQL = await initSqlJs({ locateFile: () => wasmPath });
-    const fileBuffer = await fs.readFile(dbPath);
-    const db = new SQL.Database(fileBuffer);
-    const res = db.exec(`SELECT value FROM ItemTable WHERE key = 'cursorAuth/cachedEmail';`);
-    db.close();
-
-    if (res && res.length > 0 && res[0].values && res[0].values.length > 0) {
-      const val = res[0].values[0][0];
-      return typeof val === 'string' ? val : null;
-    }
-    return null;
-  } catch (error) {
-    logWithTime(`读取 cachedEmail 失败: ${error}`);
-    return null;
-  }
-}
-
-async function getPrimarySessionToken(context: vscode.ExtensionContext): Promise<string | null> {
-  if (APP_TYPE !== 'cursor') {
-    return null;
-  }
-
-  const accessToken = await readAccessTokenFromDb(context);
-  if (!accessToken) {
-    return null;
-  }
-
-  return constructSessionToken(accessToken);
-}
-
-// ==================== Trae 使用量统计类型 ====================
-interface TraeUsageStats {
-  totalUsage: number;
-  totalLimit: number;
-  hasValidPacks: boolean;
-}
-
-// 副账号数据接口
-interface SecondaryAccountData {
-  summary: UsageSummaryResponse;
-  billingCycle: BillingCycleResponse | null;
-  aggregatedData: AggregatedUsageResponse | null;
-}
-
-// ==================== 主类 ====================
-export class CodingUsageProvider {
-  private billingCycleData: BillingCycleResponse | null = null;
-  private summaryData: UsageSummaryResponse | null = null;
-  private aggregatedUsageData: AggregatedUsageResponse | null = null;
-  private traeUsageData: TraeApiResponse | null = null;
-  private secondaryAccountsData: Map<string, SecondaryAccountData> = new Map();
-  private primaryEmail: string | null = null;
-  private retryTimer: NodeJS.Timeout | null = null;
-  private clickTimer: NodeJS.Timeout | null = null;
-  private fetchTimeoutTimer: NodeJS.Timeout | null = null;
-  private statusBarItem: vscode.StatusBarItem;
-  private apiService = getApiService();
-  private clickCount = 0;
-  private isRefreshing = false;
-  private isManualRefresh = false;
-  private isAuthFailed = false;
-
-  constructor(private context: vscode.ExtensionContext) {
-    this.statusBarItem = this.createStatusBarItem();
-    this.initialize();
-  }
-
-  public showOutput(): void {
-    const outputChannel = getOutputChannel();
-    outputChannel.show();
-  }
-
-  private createStatusBarItem(): vscode.StatusBarItem {
-    const item = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 100);
-    item.command = 'cursorUsage.handleStatusBarClick';
-    item.show();
-    return item;
-  }
-
-  private initialize(): void {
-    // 主账号从 DB 实时获取，无需检查配置
-    if (APP_TYPE === 'cursor') {
-      this.isRefreshing = true;
-      this.setLoadingState();
-    } else {
-      this.updateStatusBar();
-    }
-
-    this.fetchData();
-  }
-
-  // ==================== 点击处理 ====================
-  handleStatusBarClick(): void {
-    if (this.isRefreshing) {
-      logWithTime('当前正在刷新中，忽略点击');
-      return;
-    }
-
-    this.clickCount++;
-
-    if (this.clickTimer) {
-      this.clearClickTimer();
-      vscode.commands.executeCommand('cursorUsage.updateSession');
-    } else {
-      this.clickTimer = setTimeout(() => {
-        if (this.clickCount === 1) {
-          this.refresh();
-        }
-        this.clearClickTimer();
-      }, DOUBLE_CLICK_DELAY);
-    }
-  }
-
-  private clearClickTimer(): void {
-    if (this.clickTimer) {
-      clearTimeout(this.clickTimer);
-      this.clickTimer = null;
-    }
-    this.clickCount = 0;
-  }
-
-  // ==================== 状态检查和恢复 ====================
-  public isInRefreshingState(): boolean {
-    return this.isRefreshing;
-  }
-
-  public safeRefresh(): void {
-    if (this.isRefreshing) {
-      logWithTime('重置可能卡住的刷新状态');
-      this.resetRefreshState();
-    }
-    this.fetchData();
-  }
-
-  // ==================== 刷新逻辑 ====================
-  refresh(): void {
-    logWithTime('手动刷新开始');
-    this.isManualRefresh = true;
-    this.isRefreshing = true;
-    this.isAuthFailed = false;
-    this.setLoadingState();
-
-    // Trae 需要清除缓存
-    if (APP_TYPE === 'trae') {
-      this.apiService.clearTraeCache();
-    }
-
-    this.fetchData();
-  }
-
-  private setLoadingState(): void {
-    this.statusBarItem.text = '$(loading~spin) Loading...';
-    this.statusBarItem.tooltip = 'Refreshing usage data...';
-    this.statusBarItem.color = undefined;
-  }
-
-  // ==================== 状态栏更新 ====================
-  private updateStatusBar(): void {
-    if (this.isRefreshing) {
-      this.setLoadingState();
-      return;
-    }
-
-    if (this.isAuthFailed) {
-      this.showAuthFailedStatus();
-      return;
-    }
-
-    if (APP_TYPE === 'cursor') {
-      this.updateCursorStatusBar();
-    } else if (APP_TYPE === 'trae') {
-      this.updateTraeStatusBar();
-    } else {
-      this.showNotConfiguredStatus();
-    }
-  }
-
-  private updateCursorStatusBar(): void {
-    if (!this.summaryData || !this.billingCycleData) {
-      return;
-    }
-    this.showCursorUsageStatus();
-  }
-
-  private updateTraeStatusBar(): void {
-    if (!this.traeUsageData || this.traeUsageData.code === 1001) {
-      return;
-    }
-
-    const stats = this.calculateTraeUsageStats();
-    if (stats.hasValidPacks) {
-      this.showTraeUsageStatus(stats);
-    } else {
-      this.showNoActiveSubscriptionStatus();
-    }
-  }
-
-  private showNotConfiguredStatus(): void {
-    this.statusBarItem.text = `$(warning) ${APP_NAME}: Not Configured`;
-    this.statusBarItem.color = undefined;
-    this.statusBarItem.tooltip = 'Click to configure session token\n\nSingle click: Refresh\nDouble click: Configure';
-  }
-
-  private showAuthFailedStatus(): void {
-    this.statusBarItem.text = `⚠️ ${APP_NAME}: Auth Failed`;
-    this.statusBarItem.color = '#ff6b6b';
-    this.statusBarItem.tooltip = 'Authentication failed: Session may be invalid or expired\nClick to reconfigure\n\nSingle click: Refresh\nDouble click: Configure';
-  }
-
-  private showNoActiveSubscriptionStatus(): void {
-    this.statusBarItem.text = `$(info) ${APP_NAME}: No Active Subscription`;
-    this.statusBarItem.color = undefined;
-    this.statusBarItem.tooltip = 'No active subscription pack found\n\nSingle click: Refresh\nDouble click: Configure';
-  }
-
-  // ==================== Cursor 状态显示 ====================
-  private showCursorUsageStatus(): void {
-    if (!this.summaryData || !this.billingCycleData) return;
-
-    const membershipType = this.summaryData.membershipType.toUpperCase();
-    const plan = this.summaryData.individualUsage.plan;
-
-    // 使用新的百分比字段
-    const apiPercentUsed = plan.apiPercentUsed ?? 0;
-    const autoPercentUsed = plan.autoPercentUsed ?? 0;
-    const totalPercentUsed = plan.totalPercentUsed ?? 0;
-
-    // 从聚合数据计算 API 和 Auto 使用量（美分）
-    const { apiUsageCents, autoUsageCents } = this.calculateUsageFromAggregated();
-
-    // 反推限额（如果百分比 > 0）
-    const apiLimitCents = apiPercentUsed > 0 ? (apiUsageCents / apiPercentUsed) * 100 : 0;
-
-    if (apiPercentUsed > 0 || autoPercentUsed > 0) {
-      // 显示 API 使用进度
-      const apiUsageDollars = apiUsageCents / 100;
-      const apiLimitDollars = apiLimitCents / 100;
-      this.statusBarItem.text = `⚡ ${membershipType}: $${apiUsageDollars.toFixed(2)}/${apiLimitDollars.toFixed(0)} (${apiPercentUsed.toFixed(1)}%)`;
-    } else {
-      // 回退到总体使用量显示
-      const usedCents = plan.breakdown?.total ?? plan.used;
-      const usedDollars = usedCents / 100;
-      const limitDollars = plan.limit / 100;
-      this.statusBarItem.text = `⚡ ${membershipType}: $${usedDollars.toFixed(2)}/${limitDollars.toFixed(0)} (${totalPercentUsed.toFixed(1)}%)`;
-    }
-
-    this.statusBarItem.color = undefined;
-    this.statusBarItem.tooltip = this.buildCursorDetailedTooltip();
-  }
-
-  /**
-   * 从聚合数据计算 API 和 Auto 使用量
-   */
-  private calculateUsageFromAggregated(): { apiUsageCents: number; autoUsageCents: number } {
-    if (!this.aggregatedUsageData) {
-      return { apiUsageCents: 0, autoUsageCents: 0 };
-    }
-
-    let apiUsageCents = 0;
-    let autoUsageCents = 0;
-
-    for (const event of this.aggregatedUsageData.aggregations) {
-      if (event.modelIntent === 'default') {
-        autoUsageCents += event.totalCents;
-      } else {
-        apiUsageCents += event.totalCents;
-      }
-    }
-
-    return { apiUsageCents, autoUsageCents };
-  }
-
-  // ==================== Trae 状态显示 ====================
-  private showTraeUsageStatus(stats: TraeUsageStats): void {
-    const { totalUsage, totalLimit } = stats;
-    const remaining = totalLimit - totalUsage;
-    const remainingFormatted = remaining.toFixed(1);
-
-    this.statusBarItem.text = `⚡ Fast: ${totalUsage}/${totalLimit} (${remainingFormatted} Left)`;
-    this.statusBarItem.color = undefined;
-    this.statusBarItem.tooltip = this.buildTraeDetailedTooltip();
-  }
-
-  private calculateTraeUsageStats(): TraeUsageStats {
-    let totalUsage = 0;
-    let totalLimit = 0;
-    let hasValidPacks = false;
-
-    if (!this.traeUsageData) {
-      return { totalUsage, totalLimit, hasValidPacks };
-    }
-
-    this.traeUsageData.user_entitlement_pack_list.forEach(pack => {
-      const usage = pack.usage.premium_model_fast_amount;
-      const limit = pack.entitlement_base_info.quota.premium_model_fast_request_limit;
-
-      if (limit > 0) {
-        totalUsage += usage;
-        totalLimit += limit;
-        hasValidPacks = true;
-      }
-    });
-
-    return { totalUsage, totalLimit, hasValidPacks };
-  }
-
-  // ==================== Cursor Tooltip 构建 ====================
-  private buildCursorDetailedTooltip(): vscode.MarkdownString {
-    return CodingUsageProvider.buildCursorTooltipFromData(
-      this.summaryData,
-      this.billingCycleData,
-      this.aggregatedUsageData,
-      this.secondaryAccountsData,
-      this.primaryEmail,
-      new Date()
-    );
-  }
-
-  public static buildCursorTooltipFromData(
-    summary: UsageSummaryResponse | null,
-    billing: BillingCycleResponse | null,
-    aggregatedData: AggregatedUsageResponse | null,
-    secondaryAccounts?: Map<string, SecondaryAccountData>,
-    primaryEmail?: string | null,
-    currentTime?: Date
-  ): vscode.MarkdownString {
-    if (!summary || !billing) {
-      return new vscode.MarkdownString('Primary account not detected. Please ensure you are logged in to Cursor.\n\nSingle click: Refresh\nDouble click: Settings');
-    }
-
-    const membershipType = summary.membershipType.toUpperCase();
-    const label = CodingUsageProvider.getCursorSubscriptionTypeLabel(membershipType);
-    const plan = summary.individualUsage.plan;
-    const startTime = formatTimeWithoutYear(Number(billing.startDateEpochMillis));
-    const endTime = formatTimeWithoutYear(Number(billing.endDateEpochMillis));
-    const billingPeriod = `${startTime}-${endTime}`;
-
-    const md = new vscode.MarkdownString();
-    md.supportHtml = true;
-    md.isTrusted = true;
-
-    // 从聚合数据计算使用量
-    const { apiUsageCents, autoUsageCents } = CodingUsageProvider.calculateUsageFromAggregatedStatic(aggregatedData);
-
-    // 获取百分比
-    const apiPercentUsed = plan.apiPercentUsed ?? 0;
-    const autoPercentUsed = plan.autoPercentUsed ?? 0;
-    const totalPercentUsed = plan.totalPercentUsed ?? 0;
-
-    // 反推限额
-    const apiLimitCents = apiPercentUsed > 0 ? (apiUsageCents / apiPercentUsed) * 100 : 0;
-    const autoLimitCents = autoPercentUsed > 0 ? (autoUsageCents / autoPercentUsed) * 100 : 0;
-
-    // 构建第一行：标签 + 账单周期 + 连接状态 + 更新时间
-    const hintText = TeamServerClient.isTeamHintActive() ? "✅Connect " : "";
-    const now = currentTime || new Date();
-    const mm = (now.getMonth() + 1).toString().padStart(2, '0');
-    const dd = now.getDate().toString().padStart(2, '0');
-    const hh = now.getHours().toString().padStart(2, '0');
-    const min = now.getMinutes().toString().padStart(2, '0');
-    const updateTime = `🕐${mm}/${dd} ${hh}:${min}`;
-
-    // API 使用进度
-    if (apiPercentUsed > 0) {
-      const apiUsageDollars = apiUsageCents / 100;
-      const apiLimitDollars = apiLimitCents / 100;
-      const apiProgressInfo = CodingUsageProvider.buildProgressBarFromPercent(apiPercentUsed);
-
-      md.appendMarkdown(`${label}  📅${billingPeriod}\u00A0\u00A0${hintText}${updateTime}\n\n`);
-      md.appendMarkdown(`API ($${apiUsageDollars.toFixed(2)}/${apiLimitDollars.toFixed(0)}) \u00A0\u00A0\u00A0[${apiProgressInfo.progressBar}] ${apiPercentUsed.toFixed(1)}%\n`);
-    }
-
-    // Auto 使用进度
-    if (autoPercentUsed > 0) {
-      const autoUsageDollars = autoUsageCents / 100;
-      const autoLimitDollars = autoLimitCents / 100;
-      const autoProgressInfo = CodingUsageProvider.buildProgressBarFromPercent(autoPercentUsed);
-
-      md.appendMarkdown('\n');
-      md.appendMarkdown(`Auto($${autoUsageDollars.toFixed(2)}/${autoLimitDollars.toFixed(0)}) [${autoProgressInfo.progressBar}] ${autoPercentUsed.toFixed(1)}%\n`);
-    }
-
-    // 如果没有 API/Auto 数据，回退显示总体使用量
-    if (apiPercentUsed === 0 && autoPercentUsed === 0) {
-      const usedDollars = (plan.breakdown?.total ?? plan.used) / 100;
-      const limitDollars = plan.limit / 100;
-      const progressInfo = CodingUsageProvider.buildProgressBar(usedDollars, limitDollars);
-
-      md.appendMarkdown(`${label} ($${usedDollars.toFixed(2)}/${limitDollars.toFixed(0)})  📅${billingPeriod}\u00A0\u00A0${hintText}${updateTime}\n`);
-      md.appendMarkdown(`[${progressInfo.progressBar}] ${totalPercentUsed.toFixed(1)}%\n`);
-    }
-
-    // OnDemand 使用进度（如果启用）
-    const onDemand = summary.individualUsage.onDemand;
-    if (onDemand && onDemand.enabled && onDemand.limit !== null) {
-      const onDemandUsedDollars = onDemand.used / 100;
-      const onDemandLimitDollars = onDemand.limit / 100;
-      const onDemandPercent = onDemand.limit > 0 ? (onDemand.used / onDemand.limit) * 100 : 0;
-      const onDemandProgressInfo = CodingUsageProvider.buildProgressBarFromPercent(onDemandPercent);
-
-      md.appendMarkdown('\n');
-      md.appendMarkdown(`ODM ($${onDemandUsedDollars.toFixed(2)}/${onDemandLimitDollars.toFixed(0)}) [${onDemandProgressInfo.progressBar}] ${onDemandPercent.toFixed(1)}%\n`);
-    }
-
-    // Token 使用统计（放在最后）
-    if (aggregatedData && aggregatedData.aggregations && aggregatedData.aggregations.length > 0) {
-      // 构建各模型的使用量明细表格
-      const headers = ['Model', 'In', 'Out', 'Write', 'Read', 'Cost'];
-      const rows: string[][] = [];
-
-      // 按 Cost 降序排序模型
-      const sortedAggregations = [...aggregatedData.aggregations].sort((a, b) => b.totalCents - a.totalCents);
-
-      for (const agg of sortedAggregations) {
-        const modelName = CodingUsageProvider.shortenModelName(agg.modelIntent);
-        const inputTokens = parseInt(agg.inputTokens || '0');
-        const outputTokens = parseInt(agg.outputTokens || '0');
-        const cacheWriteTokens = parseInt(agg.cacheWriteTokens || '0');
-        const cacheReadTokens = parseInt(agg.cacheReadTokens || '0');
-        const costDollars = agg.totalCents / 100;
-
-        rows.push([
-          modelName,
-          CodingUsageProvider.formatTokenCount(inputTokens),
-          CodingUsageProvider.formatTokenCount(outputTokens),
-          CodingUsageProvider.formatTokenCount(cacheWriteTokens),
-          CodingUsageProvider.formatTokenCount(cacheReadTokens),
-          `$${costDollars.toFixed(2)}`
-        ]);
-      }
-
-      // 添加合计行
-      const totalInput = parseInt(aggregatedData.totalInputTokens || '0');
-      const totalOutput = parseInt(aggregatedData.totalOutputTokens || '0');
-      const totalCacheWrite = parseInt(aggregatedData.totalCacheWriteTokens || '0');
-      const totalCacheRead = parseInt(aggregatedData.totalCacheReadTokens || '0');
-      const totalCost = aggregatedData.totalCostCents / 100;
-
-      rows.push([
-        'Total',
-        CodingUsageProvider.formatTokenCount(totalInput),
-        CodingUsageProvider.formatTokenCount(totalOutput),
-        CodingUsageProvider.formatTokenCount(totalCacheWrite),
-        CodingUsageProvider.formatTokenCount(totalCacheRead),
-        `$${totalCost.toFixed(2)}`
-      ]);
-
-      md.appendMarkdown('\n');
-      md.appendCodeblock(CodingUsageProvider.generateMultiRowAsciiTable(headers, rows), 'text');
-    }
-
-    // 显示副账号使用量（样式与主账号一致）
-    if (secondaryAccounts && secondaryAccounts.size > 0) {
-      md.appendMarkdown('\n---\n');
-      md.appendMarkdown('**Additional Accounts**\n\n');
-
-      secondaryAccounts.forEach((accData, email) => {
-        const accSummary = accData.summary;
-        const accBilling = accData.billingCycle;
-        const accAggregated = accData.aggregatedData;
-        const accPlan = accSummary.individualUsage.plan;
-        const shortEmail = email.length > 25 ? email.substring(0, 22) + '...' : email;
-        const accMembership = accSummary.membershipType.toUpperCase();
-        const accLabel = CodingUsageProvider.getCursorSubscriptionTypeLabel(accMembership);
-
-        // 账单周期
-        let billingPeriod = '';
-        if (accBilling) {
-          const startTime = formatTimeWithoutYear(Number(accBilling.startDateEpochMillis));
-          const endTime = formatTimeWithoutYear(Number(accBilling.endDateEpochMillis));
-          billingPeriod = ` 📅${startTime}-${endTime}`;
-        }
-
-        // 获取百分比
-        const apiPercent = accPlan.apiPercentUsed ?? 0;
-        const autoPercent = accPlan.autoPercentUsed ?? 0;
-        const totalPercent = accPlan.totalPercentUsed ?? 0;
-
-        // 从聚合数据计算使用量（与主账号完全一致）
-        let apiUsedCents = 0;
-        let autoUsedCents = 0;
-        if (accAggregated && accAggregated.aggregations) {
-          for (const event of accAggregated.aggregations) {
-            if (event.modelIntent === 'default') {
-              autoUsedCents += event.totalCents;
-            } else {
-              apiUsedCents += event.totalCents;
-            }
-          }
-        }
-
-        // 反推限额（与主账号完全一致的公式）
-        const apiLimitCents = apiPercent > 0 ? (apiUsedCents / apiPercent) * 100 : 0;
-        const autoLimitCents = autoPercent > 0 ? (autoUsedCents / autoPercent) * 100 : 0;
-
-        md.appendMarkdown(`**${shortEmail}** (${accLabel})${billingPeriod}\n\n`);
-
-        // API 使用进度
-        if (apiPercent > 0) {
-          const apiUsageDollars = apiUsedCents / 100;
-          const apiLimitDollars = apiLimitCents / 100;
-          const apiProgressInfo = CodingUsageProvider.buildProgressBarFromPercent(apiPercent);
-          md.appendMarkdown(`API ($${apiUsageDollars.toFixed(2)}/${apiLimitDollars.toFixed(0)}) \u00A0\u00A0\u00A0[${apiProgressInfo.progressBar}] ${apiPercent.toFixed(1)}%\n\n`);
-        }
-
-        // Auto 使用进度
-        if (autoPercent > 0) {
-          const autoUsageDollars = autoUsedCents / 100;
-          const autoLimitDollars = autoLimitCents / 100;
-          const autoProgressInfo = CodingUsageProvider.buildProgressBarFromPercent(autoPercent);
-          md.appendMarkdown(`Auto($${autoUsageDollars.toFixed(2)}/${autoLimitDollars.toFixed(0)}) [${autoProgressInfo.progressBar}] ${autoPercent.toFixed(1)}%\n\n`);
-        }
-
-        // 如果没有 API/Auto，显示总体进度
-        if (apiPercent === 0 && autoPercent === 0 && totalPercent > 0) {
-          const totalUsedCents = accPlan.breakdown?.total ?? accPlan.used ?? 0;
-          const totalLimitCents = accPlan.limit ?? 0;
-          const usedDollars = totalUsedCents / 100;
-          const limitDollars = totalLimitCents / 100;
-          const totalProgressInfo = CodingUsageProvider.buildProgressBarFromPercent(totalPercent);
-          md.appendMarkdown(`Total ($${usedDollars.toFixed(2)}/${limitDollars.toFixed(0)}) [${totalProgressInfo.progressBar}] ${totalPercent.toFixed(1)}%\n\n`);
-        }
-
-        // OnDemand 进度条
-        const onDemand = accSummary.individualUsage.onDemand;
-        if (onDemand && onDemand.enabled && onDemand.limit !== null) {
-          const odmUsedDollars = onDemand.used / 100;
-          const odmLimitDollars = onDemand.limit / 100;
-          const odmPercent = onDemand.limit > 0 ? (onDemand.used / onDemand.limit) * 100 : 0;
-          const odmProgressInfo = CodingUsageProvider.buildProgressBarFromPercent(odmPercent);
-          md.appendMarkdown(`ODM ($${odmUsedDollars.toFixed(2)}/${odmLimitDollars.toFixed(0)}) [${odmProgressInfo.progressBar}] ${odmPercent.toFixed(1)}%\n\n`);
-        }
-      });
-    }
-
-    // 添加底部交互链接（实现 Sticky Hover）
-    md.appendMarkdown('\n---\n');
-    md.appendMarkdown('[Refresh](command:cursorUsage.refresh) \u00A0\u00A0 [Settings](command:cursorUsage.updateSession)');
-
-    return md;
-  }
-
-  /**
-   * 生成 ASCII 表格
-   */
-  private static generateAsciiTable(headers: string[], values: string[]): string {
-    const colWidths = headers.map((header, i) => Math.max(header.length, values[i].length) + 2);
-
-    const buildRow = (items: string[]) => {
-      return '│' + items.map((item, i) => {
-        const padding = colWidths[i] - item.length;
-        const leftPad = Math.floor(padding / 2);
-        const rightPad = padding - leftPad;
-        return ' '.repeat(leftPad) + item + ' '.repeat(rightPad);
-      }).join('│') + '│';
-    };
-
-    const buildSeparator = (start: string, mid: string, end: string, line: string) => {
-      return start + colWidths.map(w => line.repeat(w)).join(mid) + end;
-    };
-
-    const top = buildSeparator('┌', '┬', '┐', '─');
-    const middle = buildSeparator('├', '┼', '┤', '─');
-    const bottom = buildSeparator('└', '┴', '┘', '─');
-
-    return [
-      top,
-      buildRow(headers),
-      middle,
-      buildRow(values),
-      bottom
-    ].join('\n');
-  }
-
-  /**
-   * 生成多行 ASCII 表格
-   */
-  private static generateMultiRowAsciiTable(headers: string[], rows: string[][]): string {
-    // 计算每列的最大宽度
-    const colWidths = headers.map((header, colIndex) => {
-      const maxRowWidth = Math.max(...rows.map(row => (row[colIndex] || '').length));
-      return Math.max(header.length, maxRowWidth) + 2;
-    });
-
-    const buildRow = (items: string[]) => {
-      return '│' + items.map((item, i) => {
-        const padding = colWidths[i] - item.length;
-        const leftPad = Math.floor(padding / 2);
-        const rightPad = padding - leftPad;
-        return ' '.repeat(leftPad) + item + ' '.repeat(rightPad);
-      }).join('│') + '│';
-    };
-
-    const buildSeparator = (start: string, mid: string, end: string, line: string) => {
-      return start + colWidths.map(w => line.repeat(w)).join(mid) + end;
-    };
-
-    const top = buildSeparator('┌', '┬', '┐', '─');
-    const headerSep = buildSeparator('├', '┼', '┤', '─');
-    const bottom = buildSeparator('└', '┴', '┘', '─');
-
-    const result = [top, buildRow(headers), headerSep];
-    rows.forEach(row => {
-      result.push(buildRow(row));
-    });
-    result.push(bottom);
-
-    return result.join('\n');
-  }
-
-  /**
-   * 简化模型名称显示
-   */
-  public static shortenModelName(modelIntent: string): string {
-    // 简化常见模型名称
-    const mappings: Record<string, string> = {
-      'claude-4.5-opus-high-thinking': 'opus-4.5',
-      'claude-4.5-sonnet-thinking': 'sonnet-4.5',
-      'claude-4-opus-thinking': 'opus-4',
-      'claude-4-sonnet-thinking': 'sonnet-4',
-      'claude-3.5-sonnet': 'sonnet-3.5',
-      'claude-3-5-sonnet': 'sonnet-3.5',
-      'claude-3-opus': 'opus-3',
-      'gpt-5.2': 'gpt-5.2',
-      'gpt-4-turbo': 'gpt-4t',
-      'gpt-4o': 'gpt-4o',
-      'gpt-4o-mini': 'gpt-4o-m',
-      'default': 'auto'
-    };
-
-    if (mappings[modelIntent]) {
-      return mappings[modelIntent];
-    }
-
-    // 如果名称太长，进行截断
-    if (modelIntent.length > 12) {
-      return modelIntent.substring(0, 10) + '..';
-    }
-
-    return modelIntent;
-  }
-
-
-  /**
-   * 静态方法：从聚合数据计算使用量
-   */
-  public static calculateUsageFromAggregatedStatic(aggregatedData: AggregatedUsageResponse | null): { apiUsageCents: number; autoUsageCents: number } {
-    if (!aggregatedData) {
-      return { apiUsageCents: 0, autoUsageCents: 0 };
-    }
-
-    let apiUsageCents = 0;
-    let autoUsageCents = 0;
-
-    for (const event of aggregatedData.aggregations) {
-      if (event.modelIntent === 'default') {
-        autoUsageCents += event.totalCents;
-      } else {
-        apiUsageCents += event.totalCents;
-      }
-    }
-
-    return { apiUsageCents, autoUsageCents };
-  }
-
-  /**
-   * 格式化 Token 数量（K/M 单位）
-   */
-  public static formatTokenCount(count: number): string {
-    if (count >= 1000000) {
-      return `${(count / 1000000).toFixed(2)}M`;
-    } else if (count >= 1000) {
-      return `${(count / 1000).toFixed(1)}K`;
-    }
-    return String(count);
-  }
-
-  /**
-   * 根据百分比构建进度条
-   */
-  public static buildProgressBarFromPercent(percent: number): { progressBar: string; percentage: number } {
-    const progressBarLength = 30;
-    const filledLength = Math.round((percent / 100) * progressBarLength);
-    const clampedFilled = Math.max(0, Math.min(filledLength, progressBarLength));
-    const progressBar = '█'.repeat(clampedFilled) + '░'.repeat(progressBarLength - clampedFilled);
-
-    return { progressBar, percentage: Math.round(percent) };
-  }
-
-  public static getCursorSubscriptionTypeLabel(membershipType: string): string {
-    switch (membershipType.toUpperCase()) {
-      case 'PRO':
-        return 'Pro Plan';
-      case 'ULTRA':
-        return 'Ultra Plan';
-      default:
-        return membershipType || 'Unknown';
-    }
-  }
-
-  // ==================== Trae Tooltip 构建 ====================
-  private buildTraeDetailedTooltip(): vscode.MarkdownString {
-    return CodingUsageProvider.buildTraeTooltipFromData(this.traeUsageData, new Date());
-  }
-
-  public static buildTraeTooltipFromData(usageData: TraeApiResponse | null, currentTime?: Date): vscode.MarkdownString {
-    const md = new vscode.MarkdownString();
-    md.isTrusted = true;
-
-    if (!usageData || usageData.code === 1001) {
-      md.appendMarkdown('Click to configure Session ID\n\nSingle click: Refresh\nDouble click: Configure');
-      return md;
-    }
-
-    const sections: string[] = [];
-    const validPacks = CodingUsageProvider.getTraeValidPacks(usageData.user_entitlement_pack_list);
-
-    if (validPacks.length === 0) {
-      sections.push('No valid subscription packs');
-    } else {
-      const packSections = CodingUsageProvider.buildTraePackSections(validPacks);
-      sections.push(...packSections);
-    }
-
-    const hintText = TeamServerClient.isTeamHintActive() ? "✅Connected" : undefined;
-    sections.push('');
-    sections.push(CodingUsageProvider.buildTimeSection(currentTime, hintText));
-
-    md.appendMarkdown(sections.join('\n'));
-
-    // 添加底部交互链接（实现 Sticky Hover）
-    md.appendMarkdown('\n---\n');
-    md.appendMarkdown('[Refresh](command:cursorUsage.refresh) \u00A0\u00A0 [Settings](command:cursorUsage.updateSession)');
-
-    return md;
-  }
-
-  public static getTraeValidPacks(packList: TraeEntitlementPack[]): TraeEntitlementPack[] {
-    return packList.filter(pack => CodingUsageProvider.hasTraeValidUsageData(pack));
-  }
-
-  public static hasTraeValidUsageData(pack: TraeEntitlementPack): boolean {
-    const { quota } = pack.entitlement_base_info;
-    return quota.premium_model_fast_request_limit > 0 ||
-      quota.premium_model_slow_request_limit > 0 ||
-      quota.auto_completion_limit > 0 ||
-      quota.advanced_model_request_limit > 0;
-  }
-
-  public static buildTraePackSections(validPacks: TraeEntitlementPack[]): string[] {
-    const sections: string[] = [];
-
-    validPacks.forEach((pack, index) => {
-      const { usage, entitlement_base_info } = pack;
-      const { quota } = entitlement_base_info;
-
-      const subscriptionType = CodingUsageProvider.getTraeSubscriptionTypeLabel(pack);
-      const fastUsed = usage.premium_model_fast_amount;
-      const fastLimit = quota.premium_model_fast_request_limit;
-
-      if (fastLimit > 0) {
-        const progressInfo = CodingUsageProvider.buildProgressBar(fastUsed, fastLimit);
-        const header = `${subscriptionType} (${fastUsed}/${fastLimit})  Resets: ${formatTimeWithoutYear(entitlement_base_info.end_time, true)}`;
-        sections.push(header);
-        sections.push(`[${progressInfo.progressBar}] ${progressInfo.percentage}%`);
-
-        if (index < validPacks.length - 1) {
-          sections.push('');
-        }
-      }
-    });
-
-    return sections;
-  }
-
-  public static getTraeSubscriptionTypeLabel(pack: TraeEntitlementPack): string {
-    const { entitlement_base_info } = pack;
-
-    if (entitlement_base_info.product_type !== undefined) {
-      const productType = entitlement_base_info.product_type;
-      switch (productType) {
-        case 1:
-          return 'Pro Plan';
-        case 2:
-          return 'Extra Package';
-        default:
-          return 'Unknown';
-      }
-    }
-
-    const { quota } = entitlement_base_info;
-    if (quota.premium_model_fast_request_limit === -1) {
-      return 'Unlimited';
-    } else if (quota.premium_model_fast_request_limit > 1000) {
-      return 'Premium';
-    } else {
-      return 'Basic';
-    }
-  }
-
-  // ==================== 通用工具方法 ====================
-  public static buildProgressBar(used: number, limit: number): { progressBar: string; percentage: number } {
-    const percentage = limit > 0 ? Math.round((used / limit) * 100) : 0;
-    const progressBarLength = 15;
-    const filledLength = limit > 0 ? Math.round((used / limit) * progressBarLength) : 0;
-    const clampedFilled = Math.max(0, Math.min(filledLength, progressBarLength));
-    const progressBar = '█'.repeat(clampedFilled) + '░'.repeat(progressBarLength - clampedFilled);
-
-    return { progressBar, percentage };
-  }
-
-  public static buildTimeSection(currentTime?: Date, leftText?: string): string {
-    const now = currentTime || new Date();
-    const mm = (now.getMonth() + 1).toString().padStart(2, '0');
-    const dd = now.getDate().toString().padStart(2, '0');
-    const hh = now.getHours().toString().padStart(2, '0');
-    const min = now.getMinutes().toString().padStart(2, '0');
-    const updateTime = `🕐${mm}/${dd} ${hh}:${min}`;
-    const left = leftText ? `${leftText}` : '';
-    const spaceCount = left.includes('Connected') ? 25 : 45;
-    return `${left}${' '.repeat(spaceCount)}${updateTime}`;
-  }
-
-  async fetchData(retryCount = 0): Promise<void> {
-
-    // 清除之前的超时定时器
-    this.clearFetchTimeout();
-
-    // 设置超时保护
-    this.fetchTimeoutTimer = setTimeout(() => {
-      logWithTime('fetchData 超时，强制重置状态');
-      this.resetRefreshState();
-      this.updateStatusBar();
-      if (this.isManualRefresh) {
-        vscode.window.showErrorMessage('Request timeout. Please try again.');
-      }
-    }, FETCH_TIMEOUT);
-
-    try {
-      if (APP_TYPE === 'cursor') {
-        // 主账号：从 DB 实时获取 session token
-        const primaryToken = await getPrimarySessionToken(this.context);
-        if (!primaryToken) {
-          logWithTime('无法从 DB 获取主账号 token，请确保已登录 Cursor');
-          this.handleNoSessionToken();
-          return;
-        }
-
-        // 获取主账号邮箱
-        this.primaryEmail = await readCachedEmailFromDb(this.context);
-        logWithTime(`主账号邮箱: ${this.primaryEmail}`);
-
-        // 获取主账号数据（带投递）
-        await this.fetchCursorData(primaryToken);
-
-        // 获取副账号数据（不投递，仅显示使用量）
-        await this.fetchSecondaryAccountsData();
-
-      } else if (APP_TYPE === 'trae') {
-        // Trae 暂不支持多账号，使用第一个副账号配置
-        const additionalTokens = getAdditionalSessionTokens();
-        if (additionalTokens.length > 0) {
-          await this.fetchTraeData(additionalTokens[0]);
-        } else {
-          logWithTime('Trae 需要在设置中配置 Session Token');
-          this.handleNoSessionToken();
-          return;
-        }
-      } else {
-        // 未知应用类型
-        this.handleNoSessionToken();
-        return;
-      }
-
-      this.clearFetchTimeout();
-      this.resetRefreshState();  // 先重置状态
-      this.updateStatusBar();    // 再更新状态栏
-    } catch (error) {
-      logWithTime(`fetchData 发生错误: ${error}`);
-      this.clearFetchTimeout();
-      this.handleFetchError(error, retryCount);
-    }
-  }
-
-  /**
-   * 获取副账号数据（不投递，仅显示使用量）
-   */
-  private async fetchSecondaryAccountsData(): Promise<void> {
-    const additionalTokens = getAdditionalSessionTokens();
-    this.secondaryAccountsData.clear();
-
-    for (let i = 0; i < additionalTokens.length; i++) {
-      const token = additionalTokens[i];
-      try {
-        // 获取用户信息来获取邮箱
-        const userInfo = await this.apiService.fetchCursorUserInfo(token);
-        const email = userInfo.email || `Account ${i + 2}`;
-
-        // 获取 UsageSummary
-        const summary = await this.apiService.fetchCursorUsageSummary(token);
-
-        // 获取账单周期和聚合数据（与主账号一致）
-        let billingCycle: BillingCycleResponse | null = null;
-        let aggregatedData: AggregatedUsageResponse | null = null;
-        try {
-          billingCycle = await this.apiService.fetchCursorBillingCycle(token);
-          const billingStartMillis = parseInt(billingCycle.startDateEpochMillis);
-          aggregatedData = await this.apiService.fetchCursorAggregatedUsage(token, billingStartMillis);
-          logWithTime(`获取副账号 ${email} 聚合数据成功: ${aggregatedData.aggregations.length} 条记录`);
-        } catch (aggError) {
-          logWithTime(`获取副账号 ${email} 聚合数据失败（非致命）: ${aggError}`);
-        }
-
-        this.secondaryAccountsData.set(email, { summary, billingCycle, aggregatedData });
-        logWithTime(`获取副账号 ${email} 数据成功`);
-      } catch (error) {
-        logWithTime(`获取副账号数据失败: ${error}`);
-        // 副账号失败不影响主流程
-      }
-    }
-  }
-
-
-  private clearFetchTimeout(): void {
-    if (this.fetchTimeoutTimer) {
-      clearTimeout(this.fetchTimeoutTimer);
-      this.fetchTimeoutTimer = null;
-    }
-  }
-
-  private async fetchCursorData(sessionToken: string): Promise<void> {
-    try {
-      // 1. 获取使用摘要
-      const summary = await this.apiService.fetchCursorUsageSummary(sessionToken);
-      logWithTime('获取 Cursor 使用量摘要成功');
-
-      // 从 summary 中获取账单周期时间
-      const startMillis = new Date(summary.billingCycleStart).getTime();
-      const endMillis = new Date(summary.billingCycleEnd).getTime();
-
-      this.billingCycleData = {
-        startDateEpochMillis: String(startMillis),
-        endDateEpochMillis: String(endMillis)
-      };
-      this.summaryData = summary;
-
-      // 2. 获取精确的账单周期（用于聚合数据查询）
-      try {
-        const billingCycle = await this.apiService.fetchCursorBillingCycle(sessionToken);
-        const billingStartMillis = parseInt(billingCycle.startDateEpochMillis);
-
-        // 3. 获取聚合使用数据
-        const aggregatedUsage = await this.apiService.fetchCursorAggregatedUsage(sessionToken, billingStartMillis);
-        this.aggregatedUsageData = aggregatedUsage;
-        logWithTime(`获取聚合数据成功: ${aggregatedUsage.aggregations.length} 条记录, 总费用 $${(aggregatedUsage.totalCostCents / 100).toFixed(2)}`);
-      } catch (aggError) {
-        logWithTime(`获取聚合数据失败（非致命）: ${aggError}`);
-        // 聚合数据获取失败不影响主流程
-      }
-
-      await TeamServerClient.submitCursorUsage(sessionToken, summary, this.billingCycleData, this.aggregatedUsageData);
-    } catch (error) {
-      logWithTime(`获取 Cursor 数据失败: ${error}`);
-      throw error;
-    }
-  }
-
-  private async fetchTraeData(sessionId: string): Promise<void> {
-    logWithTime('开始获取 Trae 数据');
-    try {
-      const authToken = await this.apiService.getTraeTokenFromSession(sessionId, 0, this.isManualRefresh);
-      if (!authToken) {
-        logWithTime('获取 Trae 认证令牌失败');
-        this.isAuthFailed = true;
-        return;
-      }
-
-      logWithTime('成功获取 Trae 认证令牌');
-      const responseData = await this.apiService.getTraeUserEntitlementList(authToken);
-      if (responseData) {
-        logWithTime('成功获取 Trae 使用量数据');
-        this.traeUsageData = responseData;
-        this.isAuthFailed = false;
-
-        if (!this.apiService.isApiResponseSuccess(responseData)) {
-          logWithTime(`Trae API 返回错误: code=${responseData?.code}`);
-          this.apiService.handleTraeApiResponseError(responseData, '获取使用量数据');
-          if (responseData?.code === 1001) {
-            this.isAuthFailed = true;
-            this.apiService.clearTraeCache();
-          }
-        } else {
-          await this.submitTraeDataToTeamServer(responseData);
-        }
-      } else {
-        logWithTime('Trae 返回数据为空');
-      }
-    } catch (error) {
-      logWithTime(`获取 Trae 数据失败: ${error}`);
-      throw error;
-    }
-  }
-
-  private async submitTraeDataToTeamServer(data: TraeApiResponse): Promise<void> {
-    if (!data || !data.user_entitlement_pack_list || data.user_entitlement_pack_list.length === 0) {
-      return;
-    }
-
-    let totalUsage = 0;
-    let totalLimit = 0;
-    let expireTime = 0;
-    let email = '';
-    let membershipType = 'free';
-
-    for (const pack of data.user_entitlement_pack_list) {
-      const usage = pack.usage.premium_model_fast_amount;
-      const limit = pack.entitlement_base_info.quota.premium_model_fast_request_limit;
-
-      if (limit > 0) {
-        totalUsage += usage;
-        totalLimit += limit;
-
-        if (pack.entitlement_base_info.end_time > expireTime) {
-          expireTime = pack.entitlement_base_info.end_time;
-        }
-
-        if (!email && pack.entitlement_base_info.user_id) {
-          email = pack.entitlement_base_info.user_id;
-        }
-
-        if (pack.entitlement_base_info.product_type === 1) {
-          membershipType = 'pro';
-        }
-      }
-    }
-
-    if (totalLimit > 0) {
-      await TeamServerClient.submitTraeUsage(email, {
-        expire_time: expireTime,
-        total_usage: totalLimit,
-        used_usage: totalUsage,
-        membership_type: membershipType
-      });
-    }
-  }
-
-  private resetRefreshState(): void {
-    this.isManualRefresh = false;
-    this.isRefreshing = false;
-    this.clearFetchTimeout();
-  }
-
-  // ==================== 错误处理 ====================
-  private handleNoSessionToken(): void {
-    logWithTime('处理无 session token 情况');
-    if (this.isManualRefresh) {
-      this.showSetSessionMessage();
-    }
-    this.resetRefreshState();  // 先重置
-    this.updateStatusBar();    // 再更新
-  }
-
-
-  private handleFetchError(error: any, retryCount: number): void {
-    logWithTime(`获取数据失败 (尝试 ${retryCount + 1}/${MAX_RETRY_COUNT}): ${error}`);
-
-    // 处理401认证失败情况
-    if (error.response?.status === 401) {
-      logWithTime('检测到 401 认证失败');
-      this.isAuthFailed = true;
-      this.resetRefreshState();  // 先重置
-      this.updateStatusBar();    // 再更新
-
-      if (this.isManualRefresh) {
-        vscode.window.showErrorMessage(
-          '认证失败：Session可能无效或已过期，请更新Session',
-          '更新Session'
-        ).then(selection => {
-          if (selection === '更新Session') {
-            vscode.commands.executeCommand('cursorUsage.updateSession');
-          }
-        });
-      }
-      return;
-    }
-
-    if (this.isManualRefresh) {
-      const message = isRetryableError(error)
-        ? 'Network is unstable. Please try again later.'
-        : `Failed to get usage data: ${error?.toString() || 'Unknown error'}`;
-
-      vscode.window.showErrorMessage(message);
-      this.resetRefreshState();  // 先重置
-      this.updateStatusBar();    // 再更新
-      return;
-    }
-
-    if (retryCount < MAX_RETRY_COUNT) {
-      this.scheduleRetry(retryCount);
-    } else {
-      logWithTime('API调用失败，已达到最大重试次数，停止重试');
-      this.resetRefreshState();  // 先重置
-      this.updateStatusBar();    // 再更新
-    }
-  }
-
-
-  private scheduleRetry(retryCount: number): void {
-    logWithTime(`API调用失败，将在1秒后进行第${retryCount + 1}次重试`);
-    this.retryTimer = setTimeout(() => {
-      this.fetchData(retryCount + 1);
-    }, RETRY_DELAY);
-  }
-
-  // ==================== 消息显示 ====================
-  private showSetSessionMessage(): void {
-    vscode.window.showWarningMessage(
-      'Please set your session token.',
-      'Set Token'
-    ).then(selection => {
-      if (selection === 'Set Token') {
-        vscode.commands.executeCommand('cursorUsage.updateSession');
-      }
-    });
-  }
-
-  // ==================== 清理 ====================
-  dispose(): void {
-    if (this.retryTimer) {
-      clearTimeout(this.retryTimer);
-      this.retryTimer = null;
-    }
-    this.clearClickTimer();
-    this.clearFetchTimeout();
-    if (this.statusBarItem) {
-      this.statusBarItem.dispose();
-    }
-    disposeOutputChannel();
-  }
-}
-
-// ==================== 数据库监控 ====================
-class DbMonitor {
-  private interval: NodeJS.Timeout | null = null;
-  private lastContentHash: string | null = null;
-  private wasmPath: string;
-
-  constructor(private context: vscode.ExtensionContext, private triggerRefresh: () => void) {
-    this.wasmPath = vscode.Uri.joinPath(this.context.extensionUri, 'out', 'sql-wasm.wasm').fsPath;
-  }
-
-  private async getStateDbPathForCurrentWorkspace(): Promise<string | null> {
-    const workspaceFolders = vscode.workspace.workspaceFolders;
-    if (!workspaceFolders || workspaceFolders.length === 0) {
-      return null;
-    }
-    const workspaceDir = workspaceFolders[0].uri.fsPath;
-    try {
-      if (!(await fs.pathExists(workspaceDir))) {
-        return null;
-      }
-      const stats = await fs.stat(workspaceDir);
-      const ctime = (stats as any).birthtimeMs || (stats as any).ctimeMs;
-      const normalizedPath = os.platform() === 'win32' ? workspaceDir.replace(/^([A-Z]):/, (_match, letter) => (letter as string).toLowerCase() + ':') : workspaceDir;
-      const hashInput = normalizedPath + Math.floor(ctime).toString();
-      const workspaceId = crypto.createHash('md5').update(hashInput, 'utf8').digest('hex');
-      let baseStoragePath: string;
-      const platform = os.platform();
-      const homeDir = os.homedir();
-
-      // 根据应用类型确定存储路径
-      const appFolderName = APP_TYPE === 'trae' ? 'Trae' : (vscode.env.appName || 'Cursor');
-
-      switch (platform) {
-        case 'win32': {
-          const appData = process.env.APPDATA || path.join(homeDir, 'AppData', 'Roaming');
-          baseStoragePath = path.join(appData, appFolderName, 'User', 'workspaceStorage');
-          break;
-        }
-        case 'darwin':
-          baseStoragePath = path.join(homeDir, 'Library', 'Application Support', appFolderName, 'User', 'workspaceStorage');
-          break;
-        default:
-          baseStoragePath = path.join(homeDir, '.config', appFolderName, 'User', 'workspaceStorage');
-          break;
-      }
-      const stateDbPath = path.join(baseStoragePath, workspaceId, 'state.vscdb');
-      if (await fs.pathExists(stateDbPath)) {
-        return stateDbPath;
-      }
-      return null;
-    } catch {
-      return null;
-    }
-  }
-
-  private async queryMonitoredContent(stateDbPath: string): Promise<string | null> {
-    const SQL = await initSqlJs({ locateFile: () => this.wasmPath });
-    const fileBuffer = await fs.readFile(stateDbPath);
-    const db = new SQL.Database(fileBuffer);
-    const key = getDbMonitorKey();
-    const res = db.exec(`SELECT value FROM ItemTable WHERE key = '${key}';`);
-    db.close();
-    if (res && res.length > 0 && res[0].values && res[0].values.length > 0) {
-      const val = res[0].values[0][0];
-      return typeof val === 'string' ? val : JSON.stringify(val);
-    }
-    return null;
-  }
-
-  private async tick(): Promise<void> {
-    try {
-      const dbPath = await this.getStateDbPathForCurrentWorkspace();
-      if (!dbPath) {
-        return;
-      }
-      const content = await this.queryMonitoredContent(dbPath);
-      if (!content) {
-        return;
-      }
-      const contentHash = crypto.createHash('md5').update(content, 'utf8').digest('hex');
-      if (this.lastContentHash !== contentHash) {
-        logWithTime(`[DbMonitor] 内容变化: ${this.lastContentHash?.slice(0, 8) ?? 'null'} -> ${contentHash.slice(0, 8)}`);
-        this.lastContentHash = contentHash;
-        this.triggerRefresh();
-      }
-    } catch (e: any) {
-      logWithTime(`[DbMonitor] FAILED: ${e?.message ?? e}`);
-    }
-  }
-
-  public async refresh(): Promise<void> {
-    await this.tick();
-  }
-
-  public async start(): Promise<void> {
-    const dbPath = await this.getStateDbPathForCurrentWorkspace();
-    // logWithTime(`[DbMonitor] 监控数据库路径: ${dbPath}`);
-
-    await this.tick();
-    this.interval = setInterval(() => this.tick(), 10000);
-  }
-
-  public stop(): void {
-    if (this.interval) {
-      clearInterval(this.interval);
-      this.interval = null;
-    }
-  }
-}
-
-// ==================== 剪贴板监控 ====================
-class ClipboardMonitor {
-  private lastNotifiedToken: string | null = null;
-
-  async checkForToken(): Promise<void> {
-    try {
-      const clipboardText = await vscode.env.clipboard.readText();
-      const tokenPattern = getClipboardTokenPattern();
-      const tokenMatch = clipboardText.match(tokenPattern);
-      if (tokenMatch?.[1]) {
-        await this.handleTokenDetected(tokenMatch[1]);
-      }
-    } catch (error) {
-      logWithTime(`Clipboard check failed: ${error}`);
-    }
-  }
-
-  private async handleTokenDetected(token: string): Promise<void> {
-    // 检查是否已在副账号列表中
-    const existingTokens = getAdditionalSessionTokens();
-    if (existingTokens.includes(token)) {
-      if (this.lastNotifiedToken !== token) {
-        vscode.window.showInformationMessage(`Session token already in additional accounts.`);
-        this.lastNotifiedToken = token;
-      }
-      return;
-    }
-
-    if (this.lastNotifiedToken !== token) {
-      await this.promptAddToken(token);
-    }
-  }
-
-  private async promptAddToken(token: string): Promise<void> {
-    const message = APP_TYPE === 'cursor'
-      ? 'Found session token in clipboard. Add as additional account?'
-      : 'Found session token in clipboard. Update Trae configuration?';
-
-    const choice = await vscode.window.showInformationMessage(
-      message,
-      'Add',
-      'Cancel'
-    );
-
-    if (choice === 'Add') {
-      const existingTokens = getAdditionalSessionTokens();
-      if (existingTokens.length >= 3) {
-        vscode.window.showWarningMessage('Maximum 3 additional accounts allowed. Please remove one first.');
-        return;
-      }
-
-      // 添加到副账号列表
-      const newTokens = [...existingTokens, token];
-      await getConfig().update('additionalSessionTokens', newTokens, vscode.ConfigurationTarget.Global);
-
-      this.lastNotifiedToken = token;
-      vscode.window.showInformationMessage('Session token added to additional accounts.');
-      vscode.commands.executeCommand('cursorUsage.refresh');
-    }
-  }
-}
-
-// ==================== 扩展激活 ====================
 export async function activate(context: vscode.ExtensionContext) {
-  logWithTime(`${APP_NAME} Usage Monitor extension is now activated. AppType: ${APP_TYPE}`);
+  const appType = getAppType();
+  logWithTime(`${APP_NAME} Usage Monitor extension is now activated. AppType: ${appType}`);
 
-  const provider = new CodingUsageProvider(context);
+  const providers: IUsageProvider[] = [];
+  if (appType === 'cursor') {
+    providers.push(new CursorProvider(context));
+  } else if (appType === 'trae') {
+    providers.push(new TraeProvider(context));
+  } else if (appType === 'antigravity') {
+    providers.push(new AntigravityProvider(context));
+  } else {
+    logWithTime('Unknown App Type, defaulting to Cursor logic');
+    providers.push(new CursorProvider(context));
+  }
+
   const clipboardMonitor = new ClipboardMonitor();
-  const dbMonitor = new DbMonitor(context, () => provider.fetchData());
+  const dbMonitor = new DbMonitor(context, () => providers.forEach(p => p.refresh()));
   const pingManager = new PingManager();
 
   // 启动数据库监控（每10秒检查变化）
@@ -1457,31 +45,31 @@ export async function activate(context: vscode.ExtensionContext) {
   pingManager.start();
   TeamServerClient.ping(true);
 
-  registerCommands(context, provider);
-  registerListeners(context, provider, clipboardMonitor);
+  registerCommands(context, providers);
+  registerListeners(context, providers, clipboardMonitor);
 
   context.subscriptions.push({
     dispose: () => {
       dbMonitor.stop();
       pingManager.stop();
-      provider.dispose();
+      providers.forEach(p => p.dispose());
     }
   });
 }
 
-function registerCommands(context: vscode.ExtensionContext, provider: CodingUsageProvider): void {
+function registerCommands(context: vscode.ExtensionContext, providers: IUsageProvider[]): void {
   const commands = [
     vscode.commands.registerCommand('cursorUsage.handleStatusBarClick', () => {
-      provider.handleStatusBarClick();
+      providers.forEach(p => p.handleStatusBarClick());
     }),
     vscode.commands.registerCommand('cursorUsage.refresh', () => {
-      provider.refresh();
+      providers.forEach(p => p.refresh());
     }),
     vscode.commands.registerCommand('cursorUsage.updateSession', async () => {
       await showUpdateSessionDialog(context);
     }),
     vscode.commands.registerCommand('cursorUsage.showOutput', () => {
-      provider.showOutput();
+      providers.forEach(p => p.showOutput());
     }),
     vscode.commands.registerCommand('cursorUsage.copyApiKey', async () => {
       const apiKey = getClientApiKey();
@@ -1503,14 +91,14 @@ function registerCommands(context: vscode.ExtensionContext, provider: CodingUsag
   context.subscriptions.push(...commands);
 }
 
-function registerListeners(context: vscode.ExtensionContext, provider: CodingUsageProvider, clipboardMonitor: ClipboardMonitor): void {
+function registerListeners(context: vscode.ExtensionContext, providers: IUsageProvider[], clipboardMonitor: ClipboardMonitor): void {
   const windowStateListener = vscode.window.onDidChangeWindowState(async (e) => {
     if (e.focused) {
       setTimeout(async () => {
         clipboardMonitor.checkForToken();
-        if (provider.isInRefreshingState()) {
+        if (providers.some(p => p.isInRefreshingState())) {
           logWithTime('检测到之前可能卡住的刷新状态，尝试恢复...');
-          provider.safeRefresh();
+          providers.forEach(p => p.safeRefresh());
         }
       }, 500);
     }
@@ -1598,13 +186,8 @@ async function showUpdateSessionDialog(context: vscode.ExtensionContext): Promis
         }
         break;
       case 'openSettings':
-        await vscode.commands.executeCommand('workbench.action.openSettings', '@ext:whyuds.coding-usage');
+        vscode.commands.executeCommand('workbench.action.openSettings', '@ext:whyuds.coding-usage');
         break;
     }
   }
-}
-
-export async function deactivate() {
-  logWithTime(`${APP_NAME} Usage Monitor extension is now deactivated.`);
-  await TeamServerClient.ping(false);
 }
