+++ CodingUsage/tsconfig.json
@@ -3,7 +3,9 @@
     "module": "commonjs",
     "target": "ES2020",
     "outDir": "out",
-    "lib": ["ES2020"],
+    "lib": [
+      "ES2020"
+    ],
     "sourceMap": true,
     "rootDir": "src",
     "strict": true,
@@ -13,5 +15,10 @@
     "resolveJsonModule": true,
     "allowSyntheticDefaultImports": true
   },
-  "exclude": ["node_modules", ".vscode-test"]
+  "exclude": [
+    "node_modules",
+    ".vscode-test",
+    "AntigravityQuotaWatcher",
+    "esbuild.js"
+  ]
 }
